Move question concept classes to buisness logic subpackage
