+++ src/main/java/com/example/demo/models/businesslogic/Choice.java
@@ -2,7 +2,6 @@ package com.example.demo.models.businesslogic;
 
 import com.example.demo.models.businesslogic.backend.BackendFact;
 import com.example.demo.models.entities.AnswerObject;
-import com.example.demo.models.entities.QuestionConceptChoice;
 import com.example.demo.models.entities.Response;
 
 import java.util.ArrayList;
