+++ src/main/java/com/example/demo/models/businesslogic/questionconcept/QuestionConceptOrder.java
@@ -1,5 +1,7 @@
-package com.example.demo.models.entities;
+package com.example.demo.models.businesslogic.questionconcept;
 
+import com.example.demo.models.entities.Backend;
+import com.example.demo.models.entities.Question;
 import lombok.Data;
 import lombok.NoArgsConstructor;
